fix:Scaffold using baseView for jobs and jobs detail view
