+++ client/lib/core/views/jobs_view/jobs.view.dart
@@ -1,5 +1,3 @@
-import 'package:client/core/l10n/app_l10n.dart';
-import 'package:client/core/views/common/widgets/custom_appbar.dart';
 import 'package:client/core/views/jobs_view/jobs.viewmodel.dart';
 import 'package:client/core/views/jobs_view/jobs.widgets.dart';
 import 'package:flutter/material.dart';
@@ -10,13 +8,10 @@ class JobsView extends StatelessWidget with JobsWidgets {
   final _vm = GetIt.I.get<JobsViewModel>();
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: CustomAppbar(
-        titleText: L10n.of(context)!.jobAndinternship,
-        isCheck: false,
-        isName: false,
-      ),
+    return _vm.baseScaffold(
+      context,
       body: body(context),
+      appbar: appbar(context),
     );
   }
 }
